+++ grails-app/domain/foodpaint/CustomerOrderDet.groovy
@@ -54,7 +54,7 @@ class CustomerOrderDet{
     /**
      * 訂單數量
      */
-	long qty
+	double qty
 
     static mapping = {
         importFlag  defaultValue: -1
