+++ grails-app/domain/foodpaint/InventoryDetail.groovy
@@ -15,7 +15,7 @@ class InventoryDetail {
         batch: Batch
     ]
 
-    long qty = 0
+    double qty = 0
 
     /**
      *  最近入庫日期
